#47 fix: 헤더 검색창에 검색하면 최근 검색어에 반영
