+++ components/WikiList/RecentSearch.module.scss
@@ -5,7 +5,7 @@
   width: 860px;
 }
 
-.theLatestSearch {
+.recentSearch {
   margin-top: 20px;
   .title {
     @include flexCenter;
