+++ components/WikiList/types/SearchType.ts
@@ -1,6 +1,5 @@
 interface SearchType {
   inputClassName?: string;
-  onAddKeyword: (text: string) => void;
 }
 
 export type { SearchType };
