+++ contexts/SearchProvider.tsx
@@ -1,14 +1,22 @@
-import { createContext, ReactNode, useCallback, useContext, useMemo, useState } from 'react';
+import { createContext, Dispatch, ReactNode, SetStateAction, useCallback, useContext, useMemo, useState } from 'react';
+
+interface KeywordsType {
+  id: number;
+  text: string;
+}
 
 interface SearchContextType {
   searchTerm: string;
   onSearch: (term: string) => void;
+  keywords: KeywordsType[];
+  setKeywords: Dispatch<SetStateAction<KeywordsType[]>>;
 }
 
 const SearchContext = createContext<SearchContextType | undefined>(undefined);
 
 export default function SearchProvider({ children }: { children: ReactNode }) {
   const [searchTerm, setSearchTerm] = useState<string>('');
+  const [keywords, setKeywords] = useState<KeywordsType[]>([]);
 
   //searchTerm을 input의 value값에 따라 변경
   const onSearch = useCallback((term: string) => {
@@ -19,8 +27,10 @@ export default function SearchProvider({ children }: { children: ReactNode }) {
     () => ({
       searchTerm,
       onSearch,
+      keywords,
+      setKeywords,
     }),
-    [searchTerm, onSearch],
+    [searchTerm, onSearch, keywords],
   );
 
   return <SearchContext.Provider value={contextValue}>{children}</SearchContext.Provider>;
