+++ pages/wikilist/index.tsx
@@ -2,7 +2,7 @@ import WikiCard from '@/components/WikiList/WikiCard';
 import styles from './WikiList.module.scss';
 import Image from 'next/image';
 import Spinner from '@/components/WikiList/SPinner';
-import TheLatestSearch from '@/components/WikiList/TheLatestSearch';
+import RecentSearch from '@/components/WikiList/RecentSearch';
 import { useSearch } from '@/contexts/SearchProvider';
 import { useWikiInfiniteScroll } from '@/hooks/WikiList/useWikiInfiniteScroll';
 
@@ -17,7 +17,7 @@ export default function WikiList() {
   return (
     <>
       <section className={styles.searchForm}>
-        <TheLatestSearch />
+        <RecentSearch />
         {hasSearchedProfile && (
           <p>
             <span>{searchTerm}</span>님을 총 <span>{profileCards.length}</span>명 찾았습니다.
